backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.26.0
cryptography>=41.0.0
pydub>=0.25.1
PyPDF2>=3.0.0